# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Location-first social and dating. Errors are RFC 9457 problem details with a stable `code`; see docs/08-flutter-integration.md.",
    "title" : "OneDay API",
    "version" : "1"
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/connections/{connectionId}/couple" : {
      "delete" : {
        "operationId" : "connectionLeaveCouple",
        "parameters" : [ {
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoupleView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Connection" ]
      },
      "post" : {
        "operationId" : "connectionCouple",
        "parameters" : [ {
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Optional, up to 100 characters: a retry with the same key and body replays the first response for 24 h",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoupleView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Connection" ]
      }
    },
    "/connections/{connectionId}/exit" : {
      "post" : {
        "operationId" : "connectionExit",
        "parameters" : [ {
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Optional, up to 100 characters: a retry with the same key and body replays the first response for 24 h",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "Connection" ]
      }
    },
    "/connections/{connectionId}/spark" : {
      "delete" : {
        "operationId" : "connectionWithdrawSpark",
        "parameters" : [ {
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SparkView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Connection" ]
      },
      "post" : {
        "operationId" : "connectionSpark",
        "parameters" : [ {
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Optional, up to 100 characters: a retry with the same key and body replays the first response for 24 h",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SparkView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Connection" ]
      }
    },
...
    "/moments" : {
      "post" : {
        "operationId" : "momentPublish",
        "parameters" : [ {
          "description" : "Optional, up to 100 characters: a retry with the same key and body replays the first response for 24 h",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateMomentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MomentView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "Moment" ]
      }
    },
...
    "/moments/trail" : {
      "get" : {
        "operationId" : "momentTrail",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TrailItem"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      }
    },
    "/moments/{momentId}" : {
      "delete" : {
        "operationId" : "momentDelete",
        "parameters" : [ {
          "in" : "path",
          "name" : "momentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      },
      "get" : {
        "operationId" : "momentView",
        "parameters" : [ {
          "in" : "path",
          "name" : "momentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MomentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      }
    },
    "/moments/{momentId}/keep" : {
      "delete" : {
        "operationId" : "momentUnkeep",
        "parameters" : [ {
          "in" : "path",
          "name" : "momentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      },
      "post" : {
        "operationId" : "momentKeep",
        "parameters" : [ {
          "in" : "path",
          "name" : "momentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Optional, up to 100 characters: a retry with the same key and body replays the first response for 24 h",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MomentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      }
    },
    "/moments/{momentId}/relay" : {
      "get" : {
        "operationId" : "momentRelay",
        "parameters" : [ {
          "in" : "path",
          "name" : "momentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelayView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Moment" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CoupleView" : {
        "type" : "object",
        "properties" : {
          "coupleMode" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "youConfirmed" : {
            "type" : "boolean"
          }
        }
      },
      "CreateMomentRequest" : {
        "type" : "object",
        "properties" : {
          "activityTag" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0
          },
          "caption" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "capturedLive" : {
            "type" : "boolean"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "PHOTO", "VIDEO", "TEXT" ]
          },
          "mediaRef" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "previewAllowed" : {
            "type" : "boolean"
          },
          "promptKey" : {
            "type" : "string",
            "maxLength" : 48,
            "minLength" : 0
          },
          "replyToMomentId" : {
            "type" : "string",
            "maxLength" : 36,
            "minLength" : 0
          },
          "shareScope" : {
            "type" : "string",
            "enum" : [ "FRIENDS_ONLY", "PUBLIC_DISCOVERY" ]
          }
        },
        "required" : [ "kind", "shareScope" ]
      },
...
      "MomentView" : {
        "type" : "object",
        "properties" : {
          "activityTag" : {
            "type" : "string"
          },
          "answersPrompt" : {
            "type" : "boolean"
          },
          "caption" : {
            "type" : "string"
          },
          "capturedLive" : {
            "type" : "boolean"
          },
          "firstName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "PHOTO", "VIDEO", "TEXT" ]
          },
          "layer" : {
            "type" : "string",
            "enum" : [ "AMBIENT", "FULL" ]
          },
          "mediaUrl" : {
            "type" : "string"
          },
          "postedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "previewUrl" : {
            "type" : "string"
          },
          "relayId" : {
            "type" : "string"
          },
          "relayPosition" : {
            "type" : "integer",
            "format" : "int32"
          },
          "shareScope" : {
            "type" : "string",
            "enum" : [ "FRIENDS_ONLY", "PUBLIC_DISCOVERY" ]
          }
        }
      },
...
      "RelayLink" : {
        "type" : "object",
        "properties" : {
          "activity" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "PHOTO", "VIDEO", "TEXT" ]
          },
          "liveCaptured" : {
            "type" : "boolean"
          },
          "mine" : {
            "type" : "boolean"
          },
          "momentId" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "previewUrl" : {
            "type" : "string"
          },
          "spotlit" : {
            "type" : "boolean"
          }
        }
      },
...
      "RelayView" : {
        "type" : "object",
        "properties" : {
          "activity" : {
            "type" : "string"
          },
          "length" : {
            "type" : "integer",
            "format" : "int32"
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RelayLink"
            }
          },
          "open" : {
            "type" : "boolean"
          },
          "relayId" : {
            "type" : "string"
          }
        }
      },
...
      "SparkView" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "mutualSpark" : {
            "type" : "boolean"
          },
          "youSparked" : {
            "type" : "boolean"
          }
        }
      },
...
      "TrailItem" : {
        "type" : "object",
        "properties" : {
          "activity" : {
            "type" : "string"
          },
          "answeredPrompt" : {
            "type" : "boolean"
          },
          "areaLat" : {
            "type" : "number",
            "format" : "double"
          },
          "areaLon" : {
            "type" : "number",
            "format" : "double"
          },
          "caption" : {
            "type" : "string"
          },
          "capturedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "PHOTO", "VIDEO", "TEXT" ]
          },
          "mediaUrl" : {
            "type" : "string"
          },
          "momentId" : {
            "type" : "string"
          },
          "partOfRelay" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}